requirements.txt
requests==2.28.1
environs==9.5.0
python-telegram-bot==13.15
google-cloud-storage==2.7.0